cleaning up Profile.tsx code
