+++ client/src/modules/Profile/Styles/PastReviews.module.css
@@ -1,0 +1,40 @@
+.pastHeader {
+  /* margin-top: 61.5px; */
+  margin-top: 30px;
+}
+
+.pastHeaderText {
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 24px;
+  line-height: 11px;
+  /* or 45% */
+
+  display: flex;
+  align-items: flex-end;
+  letter-spacing: -0.128928px;
+
+  /* Gray */
+
+  color: #a2a1a1;
+}
+
+@media (max-width: 480px) {
+  .pastHeaderText {
+    font-size: 1.2rem;
+  }
+
+  .pastHeader {
+    margin-left: 5vw;
+  }
+
+  .reviewCard {
+    background-color: transparent;
+    width: 100%;
+    overflow-y: auto;
+    padding-right: 0;
+    min-height: 18.5rem;
+    margin-bottom: 0rem;
+  }
+}
