+++ client/src/modules/Profile/Styles/PendingReviews.module.css
@@ -1,0 +1,51 @@
+.pendingHeader {
+  margin-top: 10px;
+}
+
+.pendingHeaderText {
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 24px;
+  line-height: 11px;
+  /* or 45% */
+
+  display: flex;
+  align-items: flex-end;
+  letter-spacing: -0.128928px;
+
+  /* Gray */
+
+  color: #a2a1a1;
+}
+
+.hidePending {
+  margin-right: 0px;
+}
+
+.hidePendingText {
+  text-align: right;
+  cursor: pointer;
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 19px;
+  /* or 104% */
+
+  letter-spacing: -0.225605px;
+  text-decoration-line: underline;
+
+  color: #0076ff;
+}
+
+@media (max-width: 480px) {
+  .reviewCard {
+    background-color: transparent;
+    width: 100%;
+    overflow-y: auto;
+    padding-right: 0;
+    min-height: 18.5rem;
+    margin-bottom: 0rem;
+  }
+}
