+++ client/src/modules/Profile/Styles/UserInfo.module.css
@@ -1,0 +1,212 @@
+.profileLeft {
+  display: flex;
+  margin-top: 42px;
+  margin-left: 87px;
+  padding: 0 !important;
+  height: 100%;
+}
+
+.profileContainer {
+  width: 360px;
+  height: 660px;
+  padding-bottom: 20px;
+  left: 87px;
+  background: #ffffff;
+  border-radius: 10px;
+}
+
+.profileInfo {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  justify-content: center;
+}
+
+.profileImage {
+  width: 140px;
+}
+
+.profileTitle {
+  margin-top: 10px;
+  margin-left: 37px;
+  margin-bottom: 20px;
+
+  text-align: left;
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 32px;
+  line-height: 156%;
+
+  letter-spacing: -0.128928px;
+
+  color: #000000;
+
+  border-radius: 19.6611px;
+}
+
+.profileVerifiedNetid {
+  margin-top: 8%;
+  margin-bottom: 25px;
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 38%;
+
+  display: flex;
+  align-items: center;
+  letter-spacing: -0.0819218px;
+  text-align: center;
+
+  color: #a2a1a1;
+}
+
+.profileUserStatisticsText {
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 24px;
+
+  width: 286px;
+  height: 53px;
+
+  display: flex;
+  align-items: flex-end;
+  letter-spacing: -0.128928px;
+
+  color: #000000;
+
+  border-radius: 19.6611px;
+  margin-bottom: 20px;
+}
+
+.profileSignOutButton {
+  margin-top: 35px;
+  text-align: center;
+
+  background: #0076ff;
+  border-radius: 10px;
+
+  width: 267px;
+  height: 41px;
+}
+
+.profileSignOutText {
+  height: 41px;
+  font-family: Source Sans Pro;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 23px;
+
+  display: flex;
+  align-items: center;
+  text-align: center;
+  justify-content: center;
+
+  color: #ffffff;
+}
+
+@media (min-width: 481px) and (max-width: 1000px) {
+  .profileLeft {
+    width: 100%;
+  }
+
+  .profileContainer {
+    width: 85%;
+    height: 80%;
+  }
+}
+
+@media (max-width: 480px) {
+  .profileLeft {
+    display: flex;
+    margin-top: 0px;
+    margin-left: 0px;
+    padding: 0 !important;
+    height: 100%;
+    width: 100%;
+  }
+
+  .profileContainer {
+    width: 100%;
+    height: 10rem;
+    padding-bottom: 20px;
+    left: 0px;
+    border-radius: 0px;
+    background-color: transparent;
+    display: flex;
+  }
+
+  .profileTitle,
+  .profileInfo {
+    display: inline-block;
+    width: 15%;
+    margin-right: 8rem;
+  }
+
+  .profileTitle {
+    margin-top: 10px;
+    margin-left: 1.7rem;
+    margin-bottom: 20px;
+
+    text-align: left;
+    font-family: -apple-system, BlinkMacSystemFont;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 1.5rem;
+    line-height: 156%;
+
+    letter-spacing: -0.128928px;
+
+    color: #000000;
+
+    border-radius: 0px;
+
+    white-space: nowrap;
+  }
+
+  .profileInfo {
+    position: relative;
+    left: 0rem;
+  }
+
+  .profileImage {
+    margin-top: 1rem;
+    margin-left: 5.5rem;
+    width: 2.5rem;
+  }
+
+  .profileVerifiedNetid {
+    position: relative;
+    left: -11.65rem;
+    top: -0.5rem;
+    font-size: 0.8rem;
+    color: #515151;
+  }
+
+  .profileUserStatisticsText {
+    display: none;
+  }
+
+  .profileUserStatistics {
+    margin-right: 3.5rem;
+    display: flex;
+    justify-content: space-evenly;
+  }
+
+  .divider {
+    display: inline;
+    content: '';
+    border-left: 1px solid #d9d9d9;
+    height: 6rem;
+    position: relative;
+    right: 0.9rem;
+  }
+
+  .profileSignOutButton {
+    display: none;
+  }
+}
